Moved react components into app/component/ and fixed dependancies broken by move.
