+++ app/components/Queue/Queue.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './Queue.css';
-import QueueSonglist from '../QueueSonglist/QueueSonglist';
+import QueueSonglist from '../QueueSonglist/QueueSonglist.js';
 
 export default class Queue extends React.Component {
   constructor(props) {
